% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSCI_min_vol.R
\name{printDate_min_vol}
\alias{printDate_min_vol}
\title{printDate_min_vol()}
\usage{
printDate_min_vol()
}
\description{
Print out the date when the MSCI min_vol data printed out
by printTable_min_vol() was updated
}

